feat(nav): rename navigation to just nav
