+++ packages/clay-nav/.yarnrc
@@ -1,0 +1,3 @@
+# Make `yarn version` produce the right commit message and tag for this package.
+version-tag-prefix "@clayui/nav@"
+version-git-message "chore: prepare @clayui/nav@%s"
