+++ packages/clay-nav/package.json
@@ -1,9 +1,9 @@
 {
-	"name": "@clayui/navigation",
+	"name": "@clayui/nav",
 	"version": "3.0.0-alpha.1",
-	"description": "ClayNavigation component",
+	"description": "ClayNav component",
 	"license": "BSD-3-Clause",
-	"repository": "https://github.com/liferay/clay/tree/master/packages/clay-navigation",
+	"repository": "https://github.com/liferay/clay/tree/master/packages/clay-nav",
 	"engines": {
 		"node": ">=0.12.0",
 		"npm": ">=3.0.0"
