+++ packages/clay-nav/src/index.tsx
@@ -4,10 +4,9 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-import {ClayBreadcrumbNav} from './Breadcrumb';
 import Nav from './Nav';
 import {ClayVerticalNav} from './Vertical';
 
-export {ClayBreadcrumbNav, ClayVerticalNav};
+export {ClayVerticalNav};
 
 export default Nav;
