--- packages/clay-navigation/.yarnrc
@@ -1,3 +1,0 @@
-# Make `yarn version` produce the right commit message and tag for this package.
-version-tag-prefix "@clayui/navigation@"
-version-git-message "chore: prepare @clayui/navigation@%s"
