+++ packages/demos/package.json
@@ -10,7 +10,7 @@
 		"@clayui/list": "*",
 		"@clayui/management-toolbar": "*",
 		"@clayui/multi-select": "*",
-		"@clayui/navigation": "*",
+		"@clayui/nav": "*",
 		"@clayui/pagination": "*",
 		"@clayui/panel": "*",
 		"classnames": "^2.2.6"
