+++ packages/demos/stories/FormPage.tsx
@@ -8,7 +8,7 @@ import ClayButton from '@clayui/button';
 const spritemap = require('@clayui/css/lib/images/icons/icons.svg');
 import ClayForm, {ClayInput} from '@clayui/form';
 import ClayMultiSelect from '@clayui/multi-select';
-import {ClayVerticalNav} from '@clayui/navigation';
+import {ClayVerticalNav} from '@clayui/nav';
 import ClayPanel from '@clayui/panel';
 import React, {useState} from 'react';
